refactor(Language Server): shorten languageServer folder to lsp. Move validator.ts into lsp folder directly
