+++ src/stepFunctions/asl/aslServer.ts
@@ -34,11 +34,11 @@ import {
 
 import { posix } from 'path'
 import * as URL from 'url'
-import { getLanguageModelCache } from '../../shared/languageServer/languageModelCache'
-import { formatError, runSafe, runSafeAsync } from '../../shared/languageServer/utils/runner'
+import { getLanguageModelCache } from '../../shared/lsp/languageModelCache'
+import { formatError, runSafe, runSafeAsync } from '../../shared/lsp/utils/runner'
 import { YAML_ASL, JSON_ASL } from '../constants/aslFormats'
 import globals from '../../shared/extensionGlobals'
-import { TextDocumentValidator } from '../../shared/languageServer/utils/validator'
+import { TextDocumentValidator } from '../../shared/lsp/validator'
 
 namespace ResultLimitReachedNotification {
     export const type: NotificationType<string, any> = new NotificationType('asl/resultLimitReached')
